+++ branches/GZDoomBuilder/Source/Core/Geometry/Angle2D.cs
@@ -39,7 +39,7 @@ namespace CodeImp.DoomBuilder.Geometry
 		// This converts doom angle to real angle
 		public static float DoomToReal(int doomangle)
 		{
-			return (float)Math.Round(Normalized(DegToRad((float)(doomangle + 90))), 4);
+			return (float)Math.Round(Normalized(DegToRad((doomangle + 90))), 4);
 		}
 
 		// This converts real angle to doom angle
